EXPERIMENTAL: attempt to fix double-paints and excessive shakiness.
DO NOT SUBMIT TO MASTER AS IS-- contains extra key listener
to facilitate testing: left or right arrow moves split pane divider
by 1 pixel.

Changes:

- Removed xOff,yOff as view params.  They were ints, computed overly
  circuitously in integer arithmentic, resulting in unnecessary quantization.
  If we change all that to floats instead, it turns out they are simply
  getWidth()/2.f and getHeight()/2.f, respectively.  So I removed these
  member vars and, where they were used, just use getWidth()/2.f
  and getHeight()/2.f directly, instead.

- I noticed that (on my Linux machine) there were always 2 paints of the
  MC4DView after any 1-pixel moton of the sliderpane divider,
  the first one of them with wrong view params.
  Why?  Because the following sequence of events happened:
  1. paint() got called with new size, *with old view factors*
  2. the componentResized event happened, triggering our listener
     which calls updateViewFactors(), which calls repaint()
  3. and so paint got called again, *with new view factors*.
  I think the problem was that we assumed the componentResized event
  always happens before the associated repaint.  And, in a reasonable world,
  that would probably be the case, but evidently it doesn't always happen
  that way, at least on Linux.
  So, I removed the apparently-uselesss componentResized listener,
  and changed paint() so that it starts by calling
  updateViewParams(shouldCallRepaint=false)  (everyone else calls it
  with shouldCallRepaint=true).
  Further detail: updateViewParams() is expensive (it computes a frame),
  so I made it so paint() calls it only if the view size has changed
  in a way that makes the view params stale.

With both of these changes, moving the splitpane divider is no longer shaky
for me, in both normal L&F or synthetica.
However, resizing the window from left or top is still shaky.
This is a common problem, for which I posted this question:
https://stackoverflow.com/questions/64084888/shakiness-when-resizing-a-window-by-dragging-left-or-top-border

+++ src/com/superliminal/magiccube4d/PuzzleManager.java
@@ -426,8 +426,8 @@ public class PuzzleManager
         float eyeW,
         float eyeZ,
         float viewScale,
-        int xOff,
-        int yOff,
+        float xOff,
+        float yOff,
         float towardsSunVec[], // used if showShadows is true
         boolean showShadows,
         boolean do3dStepsOnly,
